* refactor on package name psycopg insteead of psycopg3
