+++ docs-requirements.txt
@@ -36,7 +36,7 @@ kafka-python>=2.0,<3.0
 mysql-connector-python~=8.0
 mysqlclient~=2.1.1
 psutil>=5
-psycopg>=3.1.17
+psycopg~=3.1.17
 pika>=0.12.0
 pymongo~=3.1
 PyMySQL~=0.9.3
