--- docs/instrumentation/psycopg3/psycopg3.rst
@@ -1,7 +1,0 @@
-OpenTelemetry Psycopg3 Instrumentation
-=====================================
-
-.. automodule:: opentelemetry.instrumentation.psycopg3
-    :members:
-    :undoc-members:
-    :show-inheritance:
