--- instrumentation/opentelemetry-instrumentation-psycopg3/README.rst
@@ -1,21 +1,0 @@
-OpenTelemetry Psycopg Instrumentation
-=====================================
-
-|pypi|
-
-.. |pypi| image:: https://badge.fury.io/py/opentelemetry-instrumentation-psycopg3.svg
-   :target: https://pypi.org/project/opentelemetry-instrumentation-psycopg3/
-
-Installation
-------------
-
-::
-
-    pip install opentelemetry-instrumentation-psycopg3
-
-
-References
-----------
-* `OpenTelemetry Psycopg Instrumentation <https://opentelemetry-python-contrib.readthedocs.io/en/latest/instrumentation/psycopg3/psycopg3.html>`_
-* `OpenTelemetry Project <https://opentelemetry.io/>`_
-* `OpenTelemetry Python Examples <https://github.com/open-telemetry/opentelemetry-python/tree/main/docs/examples>`_
