--- instrumentation/opentelemetry-instrumentation-psycopg3/pyproject.toml
@@ -1,58 +1,0 @@
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[project]
-name = "opentelemetry-instrumentation-psycopg3"
-dynamic = ["version"]
-description = "OpenTelemetry psycopg3 instrumentation"
-readme = "README.rst"
-license = "Apache-2.0"
-requires-python = ">=3.7"
-authors = [
-  { name = "OpenTelemetry Authors", email = "cncf-opentelemetry-contributors@lists.cncf.io" },
-]
-classifiers = [
-  "Development Status :: 4 - Beta",
-  "Intended Audience :: Developers",
-  "License :: OSI Approved :: Apache Software License",
-  "Programming Language :: Python",
-  "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.7",
-  "Programming Language :: Python :: 3.8",
-  "Programming Language :: Python :: 3.9",
-  "Programming Language :: Python :: 3.10",
-  "Programming Language :: Python :: 3.11",
-]
-dependencies = [
-  "opentelemetry-api ~= 1.12",
-  "opentelemetry-instrumentation == 0.44b0.dev",
-  "opentelemetry-instrumentation-dbapi == 0.44b0.dev",
-]
-
-[project.optional-dependencies]
-instruments = [
-  "psycopg >= 3.1.17",
-]
-test = [
-  "opentelemetry-instrumentation-psycopg3[instruments]",
-  "opentelemetry-test-utils == 0.44b0.dev",
-]
-
-[project.entry-points.opentelemetry_instrumentor]
-psycopg3 = "opentelemetry.instrumentation.psycopg3:Psycopg3Instrumentor"
-
-[project.urls]
-Homepage = "https://github.com/open-telemetry/opentelemetry-python-contrib/tree/main/instrumentation/opentelemetry-instrumentation-psycopg3"
-
-[tool.hatch.version]
-path = "src/opentelemetry/instrumentation/psycopg3/version.py"
-
-[tool.hatch.build.targets.sdist]
-include = [
-  "/src",
-  "/tests",
-]
-
-[tool.hatch.build.targets.wheel]
-packages = ["src/opentelemetry"]
